Модуляция файла main js
