+++ js/generatePostsFunc.js
@@ -1,0 +1,4 @@
+import {getPost} from './js/generatePostsFunc.js';
+
+const generatePosts = () => Array.from({length:25}, getPost);
+export{generatePosts};
